Added data to url params.
The current unit and language settings only used to be tracked in the
services. However, if I wanted to copy and send a link or if I refreshed
the page, the data in my service would be reset. I thought a reasonable
solution to this would be to track the data in the url parameters.
Alternatively, I could keep track of this data in my Spring server, but
I decided this would be overkill for my purposes.